+++ jsonschema-generator/src/test/java/com/github/victools/jsonschema/generator/SchemaGeneratorTest.java
@@ -28,6 +28,7 @@ import org.junit.runners.Suite;
     SchemaGeneratorComplexTypesTest.class,
     SchemaGeneratorSubtypesTest.class,
     SchemaGeneratorCustomDefinitionsTest.class,
+    SchemaGeneratorOptionsTest.class,
     SchemaGeneratorAllOfCleanUpTest.class
 })
 public class SchemaGeneratorTest {