+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Added
+- New `Option.PROPERTIES_IN_DECLARATION_ORDER` as alternative to the default alphabetic sorting of fields and methods
 
 ## [4.11.0] - 2020-04-28
 ### `jsonschema-generator`